+++ NEWS.md
@@ -1,6 +1,7 @@
-# usedthese 0.2.0.9000
+# usedthese 0.2.0.9000 (Dev Version)
 
-* Update for dplyr 1.1 and tidyr 1.3
+* Respects `include.only` and `exclude` arguments specified in `library()`
+* Small performance improvement with dplyr 1.1 and tidyr 1.3
 
 # usedthese 0.2.0